CAMS-167: Moved detail use case code into the chapter 15 list use case and renamed that use case to chapter-15.case.ts
