--- backend/functions/lib/use-cases/chapter-15.case-detail.ts
@@ -1,30 +1,0 @@
-import { ApplicationContext } from '../adapters/types/basic';
-import { CaseDetailsDbResult } from '../adapters/types/cases';
-import { getCasesGateway } from '../factory';
-import { CasesInterface } from './cases.interface';
-
-export default class Chapter15CaseDetail {
-  casesGateway: CasesInterface;
-
-  constructor(casesGateway?: CasesInterface) {
-    if (!casesGateway) {
-      this.casesGateway = getCasesGateway();
-    } else {
-      this.casesGateway = casesGateway;
-    }
-  }
-
-  async getChapter15CaseDetail(
-    context: ApplicationContext,
-    caseId: string,
-  ): Promise<CaseDetailsDbResult> {
-    const caseDetails = await this.casesGateway.getChapter15Case(context, caseId);
-    return {
-      success: true,
-      message: '',
-      body: {
-        caseDetails,
-      },
-    };
-  }
-}
