+++ backend/functions/lib/use-cases/index.ts
@@ -1,6 +1,6 @@
 import { ApplicationContext } from '../adapters/types/basic';
 import { CaseListDbResult } from '../adapters/types/cases';
-import Chapter15CaseList from './chapter-15.case-list';
+import Chapter15CaseList from './chapter-15.case';
 import Chapter15CaseDetail from './chapter-15.case-detail';
 import AttorneysList from './attorneys';
 import InvalidChapterCaseList from './invalid-chapter.case-list';
